NFT-91 move listing mode to shared
