+++ src/nft/components/profile/list/shared.tsx
@@ -14,3 +14,10 @@ export const TitleRow = styled(Row)`
   justify-content: space-between;
   margin-bottom: 8px;
 `
+
+export enum SetPriceMethod {
+  SAME_PRICE,
+  FLOOR_PRICE,
+  LAST_PRICE,
+  CUSTOM,
+}
